enhance:1.x现在选择排序可以完善使用了2.需要完善默认的时候显示步数不正确以及距离全部都是一样的bug

+++ app/src/main/java/com/example/fazhao/locationmanager/baidu_map/model/TraceItem.java
@@ -19,16 +19,16 @@ public class TraceItem {
         this.uptime = uptime;
     }
 
-    public int getDistance() {
+    public double getDistance() {
         return distance;
     }
 
-    public void setDistance(int distance) {
+    public void setDistance(double distance) {
         this.distance = distance;
     }
 
     private int uptime = -1;
-    private int distance = -1;
+    private double distance = -1;
 
     public int getStep() {
         return step;
